// web-view/src/main/todo/src/app/angular-client/model/task.ts
/**
 * ToDo-Web-API
 * ToDo list sample app web frontend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SuperUpdateIdPageable } from './superUpdateIdPageable';


export interface Task { 
    /**
     * Task title
     */
    title: string;
    /**
     * Task descrption
     */
    description: string;
    /**
     * Task state
     */
    isDone: boolean;
    /**
     * Identifier of user that owns the task
     */
    userId?: number;
    id?: number;
    version: number;
    /**
     * Last modification date of the task
     */
    updateDate?: string;
    /**
     * Total number of pages
     */
    totalPages?: number;
    /**
     * Current page number
     */
    currentPage?: number;
}